Extract CityLayer to its own file.  It's arguably the most complex one so it's nice to have it extracted.
